+++ Button/index.js
@@ -1,15 +1,15 @@
 import React, { PropTypes } from 'react';
 import { classNames, cleanProps } from '../lib/utils';
-import styles from './Button.css';
+import styles from './button.css';
 
 const Button = (props) => {
-  const classes = classNames(styles, 'Button', {
-    'Button--state-error': props.error,
-    'Button--type-borderless': props.borderless,
-    'Button--type-secondary': props.secondary,
-    'Button--type-in-app': props.inApp,
-    'Button--type-small': props.small,
-    'Button--type-tertiary': props.tertiary,
+  const classes = classNames(styles, {
+    borderless: props.borderless,
+    large: props.large,
+    secondary: props.secondary,
+    small: props.small,
+    tertiary: props.tertiary,
+    warning: props.warning,
   });
   return (
     <button className={classes} {...cleanProps(props, Button.propTypes)}>
@@ -20,12 +20,12 @@ const Button = (props) => {
 
 Button.propTypes = {
   children: PropTypes.node,
-  error: PropTypes.bool,
   borderless: PropTypes.bool,
+  large: PropTypes.bool,
   secondary: PropTypes.bool,
-  inApp: PropTypes.bool,
   small: PropTypes.bool,
   tertiary: PropTypes.bool,
+  warning: PropTypes.bool,
 };
 
 export default Button;