+++ Card/index.js
@@ -1,13 +1,14 @@
 import React, { PropTypes } from 'react';
 import { classNames, cleanProps } from '../lib/utils';
-import styles from './Card.css';
+import styles from './card.css';
 
 const Card = (props) => {
-  const classes = classNames(styles, 'Card', {
-    'Card--double-padding': props.doublePadding,
-    'Card--empty': props.empty,
-    'Card--no-border': props.noBorder,
-    'Card--no-padding': props.noPadding,
+  const classes = classNames(styles, 'card', {
+    'double-padding': props.doublePadding,
+    empty: props.empty,
+    failed: props.failed,
+    'no-border': props.noBorder,
+    'no-padding': props.noPadding,
   });
   return (
     <div className={classes} {...cleanProps(props, Card.propTypes)}>
@@ -20,6 +21,7 @@ Card.propTypes = {
   children: PropTypes.node,
   doublePadding: PropTypes.bool,
   empty: PropTypes.bool,
+  failed: PropTypes.bool,
   noBorder: PropTypes.bool,
   noPadding: PropTypes.bool,
 };