+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 0.1.1
+
+- Fixed incompatible CSS value issue with button states
+- Added four color subclasses appended with `-light` and `-ultra-light`
+- Updated the border color of the failed `Card` configuration border
+
+## 0.1.0
+
+- Added `Loader`
+- Removed `.Button` class and applied its styles directly on the `button` node
+- Simplified CSS class names
+- Renamed `error` configuration to `warning`
+- Removed `inApp` configuration as this style is the now default style
+- Added `large` configuration
+- Simplified CSS class names
+- Added `failed` configuration
+
 ## 0.0.4
 
 - Remove `.babelrc` from npm package to fix issues